refactor: create module for server run args
