+++ zkstack_cli/crates/zkstack/src/commands/args/server/mod.rs
@@ -1,13 +1,9 @@
 use clap::{Parser, Subcommand};
-use serde::{Deserialize, Serialize};
+use run::RunServerArgs;
 
-use crate::{
-    commands::args::WaitArgs,
-    messages::{
-        MSG_SERVER_ADDITIONAL_ARGS_HELP, MSG_SERVER_COMPONENTS_HELP, MSG_SERVER_GENESIS_HELP,
-        MSG_SERVER_URING_HELP,
-    },
-};
+use crate::commands::args::WaitArgs;
+
+pub mod run;
 
 #[derive(Debug, Parser)]
 #[command(args_conflicts_with_subcommands = true, flatten_help = true)]
@@ -32,22 +28,4 @@ impl From<ServerArgs> for ServerCommand {
     fn from(args: ServerArgs) -> Self {
         args.command.unwrap_or(ServerCommand::Run(args.run))
     }
-}
-
-#[derive(Debug, Serialize, Deserialize, Parser)]
-pub struct RunServerArgs {
-    #[arg(long, help = MSG_SERVER_COMPONENTS_HELP)]
-    pub components: Option<Vec<String>>,
-    #[arg(long, help = MSG_SERVER_GENESIS_HELP)]
-    pub genesis: bool,
-    #[arg(
-        long, short,
-        trailing_var_arg = true,
-        allow_hyphen_values = true,
-        hide = false,
-        help = MSG_SERVER_ADDITIONAL_ARGS_HELP
-    )]
-    additional_args: Vec<String>,
-    #[clap(help = MSG_SERVER_URING_HELP, long, default_missing_value = "true")]
-    pub uring: bool,
 }
